+++ src/libslic3r/PrintConfig.hpp
@@ -1151,7 +1151,7 @@ PRINT_CONFIG_CLASS_DEFINE(
     ((ConfigOptionFloat,                wave_overhang_minimum_width))
     ((ConfigOptionEnum<WaveOverhangPattern>, wave_overhang_pattern))
     ((ConfigOptionFloat,                wave_overhang_line_spacing))
-    ((ConfigOptionFloat,                wave_overhang_flow_ratio))
+    ((ConfigOptionFloat,                wave_overhang_flow_mm3_per_mm))
     ((ConfigOptionFloat,                wave_overhang_print_speed))
     ((ConfigOptionFloat,                wave_overhang_travel_speed))
     ((ConfigOptionInt,                  wave_overhang_fan_speed))